# Optimize pipeline svg rendering