[add yourself]: https://github.com/concourse/governance#individual-contributors
[PR requirements]: https://github.com/concourse/concourse/blob/master/CONTRIBUTING.md#pull-request-requirements

## Changes proposed by this PR

See #7410

Applies a number of optimizations, in approximate order of impact:

* Avoid using getBBox to compute the width of graph nodes
    * Doing so requires a forced reflow - each call was taking ~2-3ms on Chrome on my machine (and we called it for every job/resource node in the graph)
    * Instead, we use the canvas `measureText` API to measure the width of text. This required a hacky fudge factor to make things consistent, but it seems to work fairly well.
    * We also hardcode the widths of material design icons and the pin icon. Should we use different images/icons in the future, this may break - but it's beneficial to avoid the reflow
* Avoid using getBBox to compute the full width/height/position of the graph
    * Instead, we compute it directly - we already know the positions/sizes of all of the nodes, so this is easy
    * This also avoids a forced reflow, but this one took ~20-30ms
* Use DOM IDs to quickly select the node we want, rather than selecting all of them and filtering by some data attributes

All these optimizations bring the render time on Chrome on https://ci.concourse-ci.org/teams/main/pipelines/concourse?group=all from 750ms (620ms of which are in rendering, esp. reflows) down to ~100ms (35ms for rendering)

**Prior to this PR**

https://user-images.githubusercontent.com/26583442/130110095-0128f321-6d1f-4029-9473-fbbb60ed2446.mov

**After this PR**

https://user-images.githubusercontent.com/26583442/130110157-27fcbc7a-f9ce-41ba-ae88-55a64f5c427b.mov


## Notes to reviewer

<!--
If needed, leave any special pointers for reviewing or testing your PR.
-->

## Release Note

* The initial render of the pipeline page should be much faster, particularly on Chrome 92+

[imperative mood]: https://chris.beams.io/posts/git-commit/#imperative